code/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

everywhereml==0.2.40
huggingface-hub==0.27.1
numpy==1.26.4
paho-mqtt==2.1.0
scikit-learn==1.6.1
tensorflow==2.16.1
tf-keras==2.16.0
transformers==4.44.0
